+++ assets/js/script.js
@@ -1,4 +1,4 @@
-import { initChart } from './grafico.js';
+import { initChart } from './mostrarGrafico.js';
 import { getData } from './consumoApi.js';
 import { paintTable } from './mostrarTabla.js';
