+++ index.html
@@ -31,7 +31,7 @@
             <div class="modal-dialog">
                 <div class="modal-content">
                     <div class="modal-header">
-                        <h5 class="modal-title" id="staticBackdropLabel">Estadísticas</h5>
+                        <h5 class="modal-title" id="staticBackdropLabel"></h5>
                         <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span>&times;</span></button>
                     </div>
                     <div class="modal-body">
@@ -66,6 +66,6 @@
         <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.0/dist/umd/popper.min.js"></script>
         <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.5.0/js/bootstrap.min.js"></script>
         <script src="https://cdn.jsdelivr.net/npm/chart.js"></script>
-        <script type="module" src="assets/scripts.js"></script>
+        <script type="module" src="assets/js/script.js"></script>
     </body>
 </html>
